+++ build.gradle
@@ -7,7 +7,7 @@ plugins {
 
 description 'Gradle file concatenation plugin'
 group = "org.jlab"
-version = "1.0.0"
+version = "1.1.0"
 
 repositories {
     mavenCentral()
