+++ src/main/java/org/jlab/cat/CatExtension.java
@@ -1,13 +1,14 @@
 package org.jlab.cat;
 
+import org.gradle.api.file.RegularFile;
 import org.gradle.api.file.RegularFileProperty;
-import org.gradle.api.provider.Property;
+import org.gradle.api.provider.ListProperty;
 import org.gradle.api.tasks.InputFiles;
 import org.gradle.api.tasks.OutputFile;
 
 public abstract class CatExtension {
     @InputFiles
-    public abstract Property<CatInput> getInput();
+    public abstract ListProperty<RegularFile> getInput();
 
     @OutputFile
     public abstract RegularFileProperty getOutput();
